Overhaul the backup, restore, and upgrade processes

+++ scripts/app-upgrade.sh
@@ -3,13 +3,19 @@ set -e
 
 START_TIME=$(date +%s)
 
+if [ ! -d .git ]; then
+    echo "[E] This script needs to run from the top directory of the repo. Current working directory:"
+    echo "      $(pwd)"
+    exit 1
+fi
+
 # The backup process will fail if the db and web containers are not started.
 
-echo "=== Starting backup container. ================================================="
-docker-compose up -d backup
+echo "=== Starting cron container. ==================================================="
+docker-compose up -d cron
 
 echo "=== Backing up application stack. =============================================="
-docker-compose exec backup app-backup
+docker-compose exec cron app-backup
 
 echo "=== Removing currnet application stack. ========================================"
 docker-compose down
@@ -24,24 +30,21 @@ echo "[I] Upgrading Bitbucket from '$OLD_BITBUCKET_VERSION' to '$NEW_BITBUCKET_V
 sed -i.bak -e "s/^BITBUCKET_VERSION=.*/BITBUCKET_VERSION=$NEW_BITBUCKET_VERSION/g" .env
 
 echo "=== Deleting old images. ======================================================="
-IMAGE_BACKUP=$(docker images ianharrier/bitbucket-backup -q)
+IMAGE_CRON=$(docker images ianharrier/bitbucket-cron -q)
 IMAGE_WEB=$(docker images ianharrier/bitbucket -q)
-docker rmi $IMAGE_BACKUP $IMAGE_WEB
+docker rmi $IMAGE_CRON $IMAGE_WEB
 
 echo "=== Building new images. ======================================================="
-docker-compose build --pull
+docker-compose build --pull --no-cache
 
 echo "=== Pulling updated database image. ============================================"
 docker-compose pull db
-
-echo "=== Starting backup container. ================================================="
-docker-compose up -d backup
 
 echo "=== Restoring application stack to most recent backup. ========================="
 cd backups
 LATEST_BACKUP=$(ls -1tr *.tar.gz 2> /dev/null | tail -n 1)
 cd ..
-docker-compose exec backup app-restore $LATEST_BACKUP
+./scripts/app-restore.sh $LATEST_BACKUP
 
 END_TIME=$(date +%s)
